+++ public/app/features/dashboard-scene/panel-edit/PanelEditNext/QueryEditor/Sidebar/Alerts/AlertCard.tsx
@@ -1,11 +1,10 @@
 import { Icon, useTheme2 } from '@grafana/ui';
 
-import { getAlertStateColor, QUERY_EDITOR_TYPE_CONFIG, QueryEditorType } from '../../constants';
-import { useQueryEditorUIContext } from '../QueryEditorContext';
-import { AlertRule } from '../types';
-
-import { CardTitle } from './CardTitle';
-import { SidebarCard } from './SidebarCard';
+import { getAlertStateColor, QUERY_EDITOR_TYPE_CONFIG, QueryEditorType } from '../../../constants';
+import { useQueryEditorUIContext } from '../../QueryEditorContext';
+import { AlertRule } from '../../types';
+import { CardTitle } from '../CardTitle';
+import { SidebarCard } from '../SidebarCard';
 
 export const AlertCard = ({ alert }: { alert: AlertRule }) => {
   const { selectedAlert, setSelectedAlert } = useQueryEditorUIContext();
